Centralize cursor token handling
