+++ src/main/java/com/amannmalik/mcp/api/ResourceOrchestrator.java
@@ -233,7 +233,7 @@ final class ResourceOrchestrator implements AutoCloseable {
     private Cursor sanitizeCursor(String cursor) {
         if (cursor == null) return Cursor.Start.INSTANCE;
         String clean = ValidationUtil.cleanNullable(cursor);
-        return new Cursor.Token(Pagination.requireValidCursor(clean));
+        return new Cursor.Token(clean);
     }
 
     private JsonRpcMessage withAccessibleUri(JsonRpcRequest req, String uri, Supplier<JsonRpcMessage> action) {
